Review of Merge and Bsort Code
The merge function is leaving some numbers behind. For instance if the two ordered arrays to merge are:
[1, 2, 3, 4] [5, 6, 7, 8]
the old code will only copy first array and will leave the other one, thus it is needed some code to join the second array

In bsort, it is necesarry to change the second for range to include all i, not i-1. Otherwise, sometimes leaves the last number of the array unordered